fix_oder item

+++ app/Http/Livewire/Admin/AdminUserComponent.php
@@ -14,9 +14,17 @@ class AdminUserComponent extends Component
     public function deleteUser()
     {
         $user=User::find($this->delete_user);
-        if (isset($user)){
-            $user->delete();
-            session()->flash('message','User has been deleted!');
+        if($user->type==0)
+        {
+            session()->flash('message',"Cannot delete admin's account!");
+        }
+        else
+        {
+            if (isset($user))
+                {
+                    $user->delete();
+                    session()->flash('message','User has been deleted!');
+                }
         } 
         redirect()->route('admin.users');
     }
